+++ signal/playerctl_v2/init.lua
@@ -1,0 +1,5 @@
+return
+{
+    cli = require(... .. ".playerctl_cli"),
+    lib = require(... .. ".playerctl_lib")
+}
